Add files via upload
New version of the program. I decided to re-define the functions that calculate the energy bands and the DOS, as well as the ones that plots them.
Now in "functions.py" there are two functions for energy band calculation (TB_1D & TB_2D) which return the energy array and the label expressing the case.
Also there are two functions for DOS calculation (DOS_1D & DOS_2D) which return the DOS, the range for plotting it, and its label.
Also there is a function called "high_symmetry_path", needed for plotting the energy band in the 2D case.

In "plots.py" I now have two similar functions (Energy_and_DOS_1D_plotter & Energy_and_DOS_2D_plotter) which plot side by side the energy band and the DOS for the 4 cases.
Also there is a function (color_map_plotter) which plot the color gradient of the energy band in the 2D cases.

In "Tight_Binding.py" I define the working parameters, I fill the list params[] which is needed by my functions, initialize the wave vectors for 1D and 2D cases and then plot the energy bands and DOS using my plotting functions.