Add files via upload
Corrected a major bug in calculations.py (new name for functions.py): TB_2D was calculating the energy band for the NNN case in a wrong way. Now it works fine!!!
Added two functions to calculations.py: "params_check()" perform a first check on each value of the list params[] with personalized error strings. "hexagonal_contour()" calculate the contour when given a bound value; I use it for plotting the FBZ in the color maps but also to define an hexagon lattice where to calculate the DOS.
I added a function called "saving_pictures()" in plot.py: it saves as pdf a plot with the same name as his title(need to be careful with illegal characters like : . , and such, but with a bit of care it's okay). Next commit I will add a feature that checks if the title of the plot contains an illegal character to raise an error value.
To finish the code I will start to do some real testing.