Add files via upload
Minor changes in the files Tight_Binding.py , plots.py , calculations.py. Added file testing.py!
testing.py contains for now testing functions on check_params(), on the rasied error messages. (unit tests)
Also testing on TB_1D() & TB_2D(). In both cases I wrote tests on the finiteness of the energy bands values, on the length of the energy band arrays, the symmetry of the bands, the correct observation of the energy boundaries, how the band should be 0 if tnn=0 and lastly how the nn and nnn case should be equal if tnnn=0.