Add files via upload
Minor changes to plot.py and Tight_Binding.py.
Added feature to calculations.py: if tnn=0, the dos is automatically equal to 0 (without normalization which would create an error).
Added tests for the two DOS functions; in both cases I test the length of the range and the dos, the finiteness of the dos and the normalization to 1. I also add two unit tests each to verify that if tnn=0 the dos is all zero, and if tnnn=0 the dos in the two cases correspond to 0.
I use hypothesis and unit tests for DOS_1D, but only unit tests for DOS_2D since, even adding the line
@settings(deadline=high time duration)
the test fails due to the illegal duration of testing... Even after removing all the assert from the test, it's still not correct. I choose to switch to unit tests with nn & nnn cases probing the two approximations with fixed values of the parameters.
The code is done! I need to check the general grammar, write the correct documentation for calculation.py and testing.py and some good commentig on plot.py and Tight_Binding.py and finish the READ_ME.txt.