Corrections on the general structure of the code.
After modifying the structure in calculations.py (both the use of the
input and the definition of the wave vectors) I changed accordingly
plots.py. At last, I modified Tight_binding.py in order to call
correctly all the various functions.
One major change in Tight_Binding.py is that now the list of parameter
can be changed by acting on an external file, called parameters.ini. The
code now read the .ini file and define the parameters for the
calculation.