Changes to the ReadMe file
CHanged the structure of the ReadMe file. Now it starts with a brief
presentation of the code. Then there is a section on how to download it
and run it. Then there is a section with the description of the
different files in the git repository. At last there is the theoretical
framework of the Tight Binding model.